Encryption stuff has been separated into more files to allow for better modularization
