+++ app/src/main/java/com/amaze/filemanager/database/typeconverters/EncryptedStringTypeConverter.java
@@ -23,6 +23,7 @@ package com.amaze.filemanager.database.typeconverters;
 import com.amaze.filemanager.application.AppConfig;
 import com.amaze.filemanager.database.models.StringWrapper;
 import com.amaze.filemanager.filesystem.files.CryptUtil;
+import com.amaze.filemanager.filesystem.files.EncryptDecrypt;
 
 import android.content.Context;
 
@@ -32,8 +33,8 @@ import androidx.room.TypeConverter;
  * {@link TypeConverter} for password strings encrypted by {@link CryptUtil}.
  *
  * @see StringWrapper
- * @see CryptUtil#encryptPassword(Context, String)
- * @see CryptUtil#decryptPassword(Context, String)
+ * @see EncryptDecrypt#encryptPassword(Context, String, String)
+ * @see EncryptDecrypt#decryptPassword(Context, String, String)
  */
 public class EncryptedStringTypeConverter {
 
@@ -43,7 +44,7 @@ public class EncryptedStringTypeConverter {
   public static StringWrapper toPassword(String encryptedStringEntryInDb) {
     try {
       return new StringWrapper(
-          CryptUtil.decryptPassword(AppConfig.getInstance(), encryptedStringEntryInDb));
+              EncryptDecrypt.decryptPassword(AppConfig.getInstance(), CryptUtil.IV, encryptedStringEntryInDb));
     } catch (Exception e) {
       android.util.Log.e(TAG, "Error decrypting password", e);
       return new StringWrapper(encryptedStringEntryInDb);
@@ -53,7 +54,7 @@ public class EncryptedStringTypeConverter {
   @TypeConverter
   public static String fromPassword(StringWrapper unencryptedPasswordString) {
     try {
-      return CryptUtil.encryptPassword(AppConfig.getInstance(), unencryptedPasswordString.value);
+      return EncryptDecrypt.encryptPassword(AppConfig.getInstance(), CryptUtil.IV, unencryptedPasswordString.value);
     } catch (Exception e) {
       android.util.Log.e(TAG, "Error encrypting password", e);
       return unencryptedPasswordString.value;
