+++ app/src/test/java/com/amaze/filemanager/ui/activities/MainActivityTest.java
@@ -55,6 +55,7 @@ import org.robolectric.util.ReflectionHelpers;
 
 import com.amaze.filemanager.application.AppConfig;
 import com.amaze.filemanager.database.UtilsHandler;
+import com.amaze.filemanager.filesystem.files.CryptUtil;
 import com.amaze.filemanager.filesystem.ssh.SshClientUtils;
 import com.amaze.filemanager.shadows.ShadowMultiDex;
 import com.amaze.filemanager.shadows.jcifs.smb.ShadowSmbFile;
@@ -197,14 +198,14 @@ public class MainActivityTest {
                 false,
                 oldName,
                 path,
-                SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), path),
+                SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), CryptUtil.IV, path),
                 null,
                 null);
             activity.addConnection(
                 true,
                 newName,
                 path,
-                SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), path),
+                SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), CryptUtil.IV, path),
                 oldName,
                 path);
