+++ app/src/test/java/com/amaze/filemanager/utils/SmbUtilTest.java
@@ -37,6 +37,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.annotation.Config;
 
+import com.amaze.filemanager.filesystem.files.CryptUtil;
 import com.amaze.filemanager.shadows.ShadowSmbAmazeFilesystem;
 import com.amaze.filemanager.shadows.ShadowSmbUtil;
 import com.amaze.filemanager.test.ShadowCryptUtil;
@@ -56,12 +57,12 @@ public class SmbUtilTest {
   public void testEncryptDecrypt() throws GeneralSecurityException, IOException {
     String path = "smb://root:toor@127.0.0.1";
     String encrypted =
-        SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), path);
+        SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), CryptUtil.IV, path);
     assertNotEquals(path, encrypted);
     assertTrue(encrypted.startsWith("smb://root:"));
     assertTrue(encrypted.endsWith("@127.0.0.1"));
     String decrypted =
-        SmbUtil.getSmbDecryptedPath(ApplicationProvider.getApplicationContext(), encrypted);
+        SmbUtil.getSmbDecryptedPath(ApplicationProvider.getApplicationContext(), CryptUtil.IV, encrypted);
     assertEquals(path, decrypted);
   }
 
@@ -69,7 +70,7 @@ public class SmbUtilTest {
   public void testEncryptWithoutCredentials() throws GeneralSecurityException, IOException {
     String path = "smb://127.0.0.1";
     assertEquals(
-        path, SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), path));
+        path, SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), CryptUtil.IV, path));
   }
 
   @Test
@@ -77,7 +78,7 @@ public class SmbUtilTest {
   public void testEncryptWithoutPassword() throws GeneralSecurityException, IOException {
     String path = "smb://toor@127.0.0.1";
     assertEquals(
-        path, SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), path));
+        path, SmbUtil.getSmbEncryptedPath(ApplicationProvider.getApplicationContext(), CryptUtil.IV, path));
   }
 
   @Test
